SDL3 event timestamps are nanoseconds

+++ app/src/ui/listwidget.c
@@ -676,7 +676,7 @@ static iBool processEvent_ListWidget_(iListWidget *d, const SDL_Event *ev) {
     if (ev->type == SDL_EVENT_MOUSE_WHEEL && isHover_Widget(w) && ev->wheel.y == 0 &&
         isPerPixel_MouseWheelEvent(&ev->wheel) && !isInertia_MouseWheelEvent(&ev->wheel)) {
         iInt2 coord = mouseCoord_SDLEvent(ev);
-        postCommand_Widget(w, "listswipe.moved arg:%d coord:%d %d", ev->wheel.x, coord.x, coord.y);
+        postCommand_Widget(w, "listswipe.moved arg:%d coord:%d %d", (int) ev->wheel.x, coord.x, coord.y);
         return iTrue;
     }
     /* Left button: item selection, dragging. */
